requirements.txt
docker
flask
ovh
python-digitalocean
tldextract
requests